requirements.txt
django<3.3
django_extensions==3.2.1
requests==2.28.1
BaseHash==3.0.10
django-tables2==2.5.0
django-debug-toolbar==3.8.1
wikipedia-api==0.5.8
django-jsoneditor==0.2.3
django_jazzmin==2.6.0
celery[redis]==5.2.7
django-celery-results==2.4.0
django-celery-beat==2.4.0
djangorestframework==3.14.0
django_ninja==0.20.0
django-bootstrap5==23.3
django-filter==23.2
uvicorn==0.23.1
gunicorn==21.2.0